Makes refactoring, comments and improves naming
