--- js/opens-photo.js
@@ -1,93 +1,0 @@
-const DISPLAY_COMMENTS = 5;
-
-const bigPictureElement = document.querySelector('.big-picture');
-const commentCountElement = bigPictureElement.querySelector('.comments-count');
-const commentListElement = bigPictureElement.querySelector('.social__comments');
-const commentsLoaderElement = bigPictureElement.querySelector('.comments-loader');
-const bodyElement = document.querySelector('body');
-const cancelButtonElement = bigPictureElement.querySelector('.big-picture__cancel');
-//Находим и получаем содержимое шаблона и записываем в переменную-объект
-const commentTemplate = document.querySelector('.big-picture__social').querySelector('.social__comment');
-
-let commentsShown = 0;
-let comments = [];
-
-//В параметрах функции деструктурируем нужные поля массива комментариев
-const createComment = ({ avatar, name, message }) => {
-  //Клонируем объект во внутреннюю переменную
-  const comment = commentTemplate.cloneNode(true);
-  //Находим поле ссылки на фото в клоне  и записываем туда реальную ссылку на аватар
-  comment.querySelector('.social__picture').src = avatar;
-  //Находим поле ссылки на имя в клоне  и записываем туда реальную ссылку на имя
-  comment.querySelector('.social__picture').alt = name;
-  comment.querySelector('.social__text').textContent = message;
-
-  return comment;
-};
-
-const renderComments = () => {
-  commentsShown += DISPLAY_COMMENTS;
-
-  if (commentsShown >= comments.length) {
-    commentsLoaderElement.classList.add('hidden');
-    commentsShown = comments.length;
-  } else {
-    commentsLoaderElement.classList.remove('hidden');
-  }
-
-  const fragment = document.createDocumentFragment();
-  for (let i = 0; i < commentsShown; i++) {
-    const comment = createComment(comments[i]);
-    fragment.append(comment);
-  }
-
-  commentListElement.innerHTML = '';
-  commentListElement.append(fragment);
-  commentCountElement.textContent = comments.length;
-};
-
-const hideBigPicture = () => {
-  bigPictureElement.classList.add('hidden');
-  bodyElement.classList.remove('modal-open');
-  document.removeEventListener('keydown', onDocumentKeydown);
-  commentsShown = 0;
-};
-
-function onDocumentKeydown(evt) {
-  if (evt.key === 'Escape') {
-    evt.preventDefault();
-    hideBigPicture();
-  }
-}
-
-const onCancelButtonClick = () => {
-  hideBigPicture();
-};
-
-const onCommentsLoaderClick = () => renderComments();
-
-const renderPicturesDetalis = ({ url, likes, description }) => {
-  bigPictureElement.querySelector('.big-picture__img img').src = url;
-  bigPictureElement.querySelector('.big-picture__img img').alt = description;
-  bigPictureElement.querySelector('.likes-count').textContent = likes;
-  bigPictureElement.querySelector('.social__caption').textContent = description;
-};
-
-const openPhoto = (data) => {
-  bigPictureElement.classList.remove('hidden');
-  bodyElement.classList.add('modal-open');
-  commentsLoaderElement.classList.add('hidden');
-  document.addEventListener('keydown', onDocumentKeydown);
-
-  renderPicturesDetalis(data);
-  comments = data.comments;
-  if (comments.length > 0) {
-    renderComments();
-  }
-};
-
-cancelButtonElement.addEventListener('click', onCancelButtonClick);
-commentsLoaderElement.addEventListener('click', onCommentsLoaderClick);
-
-
-export { openPhoto };
