+++ js/renders-mini-photos.js
@@ -1,31 +1,34 @@
-//Находим шаблон и записываем в созданную переменную
+//Модуль отрисовки фотоминиатюр на странице
+
+//Находим шаблон и записываем его расположение в переменную
 const pictureTemplate = document.querySelector('#picture');
 //Получаем содержимое шаблона и записываем в переменную-объект
 const photoTemplate = pictureTemplate.content.querySelector('.picture');
 
-//Колбэкфункция создания одного элемента - фотоминиатюры по шаблону
+//Колбэкфункция создания одного элемента, содержащиего данные о реальном фото по шаблону
+
 //В параметрах функции деструктурируем нужные поля массива описания фотографий
 const createPhoto = ({ comments, description, likes, url, id }) => {
   //Клонируем шаблон во внутреннюю переменную функции
   const photo = photoTemplate.cloneNode(true);
-
-  //Находим поле ссылки на фото в копии шаблона и записываем туда реальную ссылку на фото
-  photo.querySelector('.picture__img').src = url;
-  //Находим поле описания фото в копии шаблона и записываем туда реальное описание фото
-  photo.querySelector('.picture__img').alt = description;
   //Находим поле количества комментариев в копии шаблона и записываем туда реальную длину массива комментариев к фото
   photo.querySelector('.picture__comments').textContent = comments.length;
+  //Находим поле описания фото в копии шаблона и записываем туда реальное описание фото
+  photo.querySelector('.picture__img').alt = description;
   //Находим поле количества лайков и записываем туда реальное количество лайков к фото
   photo.querySelector('.picture__likes').textContent = likes;
-  //Создаём дата-атрибут с идентификатором фото для будушей отрисовки окна с фото в большом размере
+  //Находим поле ссылки на фото в копии шаблона и записываем туда реальную ссылку на фото
+  photo.querySelector('.picture__img').src = url;
+  //Создаём дата-атрибут с идентификатором фото для будушей отрисовки модального окна с фото в большом размере
   photo.dataset.photoIndex = id;
   //Возвращем объект - элемент фотоминиатюры
   return photo;
 };
 
-//Основная функция модуля - отрисовывает фото(добавляет массив данных о фото в заданное место DOM-дерева)
-//В параметр функции принимаем массив элементов, содержащих данные о кокретном фото
-const renderPhotos = (pictures, container) => {
+//Основная функция модуля - отрисовывает фотоминиатюры(добавляет массив данных о фото в заданное место DOM-дерева)
+
+//В параметр функции принимаем массив элементов, содержащих данные о кокретном фото и записываем в DOM
+const renderMiniPhotos = (pictures, container) => {
   //Создаём временный массив-объект для хранения фрагмента DOM-дерева с помощью метода createDocumentFragment()
   const fragment = document.createDocumentFragment();
   //Проходим по массиву элементов фото
@@ -35,8 +38,8 @@ const renderPhotos = (pictures, container) => {
     //Записывем этот элемент в конец фрагмента DOM-элементов
     fragment.append(photo);
   });
-  //Добавляем содержимое фрагмента DOM-дерева в DOM-дерево сайта
+  //Добавляем содержимое фрагмента DOM-дерева в определённое в модуле renders-page контейнер-место DOM-дерева страницы
   container.append(fragment);
 };
 
-export { renderPhotos };
+export { renderMiniPhotos };
